+++ src/app/(afterLogin)/latest/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import OptionList from "@/components/ui/option-list";
-import CardList from "@/components/ui/card-list";
+import CardList from "@/app/(afterLogin)/_component/card-list";
 import { getTrendContents } from "@/services/contents";
 import React, { useEffect, useState } from "react";
 import ClientSideLoading from "@/components/ui/client-side-loading";
@@ -47,7 +47,11 @@ function LatestContentPage() {
           onChange={onChangeOptions}
         />
       </div>
-      <div className={"grid gap-4 grid-cols-4 px-5"}>
+      <div
+        className={
+          "min-[375px]:grid-cols-2 max-w-[1480px] grid px-5 grid-col-2 gap-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5"
+        }
+      >
         <CardList dataList={data.results} inMyList={inMyList} />
       </div>
     </section>
