+++ src/components/movies/movies-info.tsx
@@ -4,7 +4,7 @@ import { getMovieContents } from "@/services/contents";
 import { useInfiniteQuery } from "react-query";
 import { useObserver } from "@/hooks/use-observer";
 import Loading from "@/app/(afterLogin)/movies/loading";
-import CardList from "@/components/ui/card-list";
+import CardList from "@/app/(afterLogin)/_component/card-list";
 import { useUserState } from "@/context/user-context";
 
 export default function MoviesInfo({ genres }: { genres: string }) {
@@ -39,14 +39,18 @@ export default function MoviesInfo({ genres }: { genres: string }) {
 
   useObserver({ target: bottom, onIntersect });
   return (
-    <>
+    <div className={"flex flex-col items-center"}>
       {status === "loading" && <Loading />}
       {status === "error" && (
         <p>에러가 발생했습니다. 잠시후 다시 시도해주세요.</p>
       )}
       {status === "success" && data && (
         <>
-          <div className={"grid gap-4 grid-cols-4 px-5"}>
+          <div
+            className={
+              "min-[375px]:grid-cols-2 max-w-[1480px] grid px-5 grid-col-2 gap-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5"
+            }
+          >
             {data?.pages.map((page) => (
               <CardList
                 key={page.page}
@@ -61,6 +65,6 @@ export default function MoviesInfo({ genres }: { genres: string }) {
       <div ref={bottom} className={"flex justify-center my-10"}>
         <span className="loading loading-spinner loading-md" />
       </div>
-    </>
+    </div>
   );
 }
